Add output_pin and sample_time to library
The variables are added to track the state globally. Probably not
thread-safe? Need to look into how Arduino execution happens, pretty
sure there are no threads.

output_pin is set to the pin we want to output a pulse onto.
sample_time controls the time between consecutive samples of the
waveform. This is how we control frequency.