add waveform generation to library
The code that generates the waveform is moved out of the loop() and made
into a function in UnityArduinoComms named generateWaveform()

Removed sign variable for changing the direction of the angle increment.

Offset sign by 127 and changed amplitude to 127 to have a full signwave
with positive and negative swing, rather than a rectified waveform.

TODO(gmicros): can we get away with a rectifier waveform, if so we can
control the amplitude, i.e., intensity in finer increments.