proj: layout the library directory structure
The structure of the library is laid out. The source needs to be copied
into the directories, independent of the sample projects.

Development and testing has been taking place in the sample projects.
That needs to stop.

There may be a needs for build scripts that copy the source into the
sample projects.